+++ services/src/main/java/org/keycloak/protocol/oidc/utils/AcrUtils.java
@@ -68,6 +68,8 @@ public class AcrUtils {
                         if (!essential || acrClaim.isEssential()) {
                             if (acrClaim.getValues() != null) {
                                 acrValues.addAll(acrClaim.getValues());
+                            } else if (acrClaim.getValue() != null) {
+                                acrValues.add(acrClaim.getValue());
                             }
                         }
                     }